Test de Diagnostico

+++ src/zork/Diagnostico.java
@@ -6,13 +6,13 @@ public class Diagnostico implements Comando{
 		float salud=jugador.getSalud();
 		String diagnostico;
 		if(salud==100) {
-			diagnostico="Tu estado de salud es perfecto ("+salud+"), solo te podria matar una seria herida";
+			diagnostico="Tu estado de salud es perfecto ("+(int)salud+"), solo te podria matar una seria herida";
 		}else if(salud<100 && salud>70) {
-			diagnostico="Tu estado de salud es bueno ("+salud+")";
+			diagnostico="Tu estado de salud es bueno ("+(int)salud+")";
 		}else if(salud<70 && salud>30) {
-			diagnostico="Tu estado de salud no es muy bueno ("+salud+"), sera mejor que encuentres algo para curarte";
+			diagnostico="Tu estado de salud no es muy bueno ("+(int)salud+"), sera mejor que encuentres algo para curarte";
 		}else {
-			diagnostico="Tu estado de salud es bajo ("+salud+"), hasta un rasguno podria ser peligroso!";
+			diagnostico="Tu estado de salud es bajo ("+(int)salud+"), hasta un rasguno podria ser peligroso!";
 		}
 		return diagnostico;
 	}
